pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teller_ai"
version = "0.2.0"
description = "TPBank LiveBank Voice Reissue — LangGraph FSM (HTTP LLM)"
authors = [{ name = "Family GPT" }]
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "PyYAML>=6.0.1",
  "typing_extensions>=4.8.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "tenacity>=9.0.0",
  "jsonschema>=4.22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.5.0",
  "mypy>=1.9.0",
  "pre-commit>=3.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
